+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support for downloading asset URLs.
+
 ### Fixed
 
 - Parsing of extract filenames that include periods (DOCSP-6904).